streams-utils/requirements.txt
contourpy==1.3.3
control==0.10.2
cycler==0.12.1
fonttools==4.59.0
h5netcdf==1.6.4
h5py==3.14.0
joblib==1.5.1
kiwisolver==1.4.9
matplotlib==3.10.5
mpi4py==4.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
Pillow==11.3.0
PyDMD==2025.8.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-sensors==0.3.5
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
wheel==0.45.1
xarray==2025.7.1
typer
PyYAML


